refactor: move collapsible and collapsible motion
