+++ packages/components/collapsible-motion/README.md
