+++ packages/components/collapsible-motion/src/index.js
