+++ packages/components/collapsible/index.js
@@ -1,3 +1,3 @@
-export { default } from '../../../src/components/collapsible';
+export { default } from './src';
 
 export { default as version } from './version';
