+++ packages/components/collapsible/src/collapsible.js
