+++ packages/components/collapsible/src/index.js
